# src/main/java/com/hulk/library/request/HandInBookRequest.java
package com.hulk.library.request;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class HandInBookRequest {
    private Long eventId;
}